[BIR] Encode BirElementBase#containingListId inside flags field
Saving one byte to be used later

+++ compiler/ir/bir/tree/src/org/jetbrains/kotlin/bir/BirChildElementList.kt
@@ -33,7 +33,7 @@ abstract class BirChildElementList<E : BirElement?>(
     }
 
     override fun contains(element: E): Boolean {
-        return element != null && element.parent === parent && (element as BirElementBase).containingListId.toInt() == id
+        return element != null && element.parent === parent && (element as BirElementBase).containingListId == id
     }
 
     override fun indexOf(element: E): Int {
@@ -80,7 +80,7 @@ abstract class BirChildElementList<E : BirElement?>(
     }
 
     protected fun BirElementBase.setContainingList() {
-        containingListId = id.toByte()
+        containingListId = id
     }
 
     protected fun BirElementBase.resetContainingList() {
